// src/app/models/buy-sale-main-model.ts
import {BuySaleModel} from './buy-sale-model';

export class BuySaleMainModel {
  data: BuySaleModel;
  employeeName?: string;
  approverName?: string;
  currencyName?: string;
  cashDeskName?: string;
  actionType?: string;
  amountFormatted?: string;
  totalAmountFormatted?: string;
  statusTr?: string;
  transactionTypeTr?: string;
  platformTr?: string;
}